+++ CHANGELOG.md
@@ -222,6 +222,7 @@
 - EXPENSE: corrected expense bank order, now if a expense bank order is cancel the expense is also cancel.
 - Expense : period / company not getting set on getOrCreateExpense method 
 - CUSTOMER : Wrong domain on action-view
+- STOCK DETAILS BY PRODUCT : remove empty mrp in progress due to stock details 
 
 ## [5.1.7] - 2019-06-17
 ## Features